+++ src/PimEnterprise/Bundle/DataGridBundle/Datagrid/ProductDraft/GridHelper.php
@@ -11,8 +11,10 @@
 
 namespace PimEnterprise\Bundle\DataGridBundle\Datagrid\ProductDraft;
 
-use Oro\Bundle\DataGridBundle\Datasource\ResultRecordInterface;
+use PimEnterprise\Bundle\SecurityBundle\Attributes;
 use PimEnterprise\Bundle\WorkflowBundle\Model\ProductDraft;
+use Oro\Bundle\DataGridBundle\Datasource\ResultRecordInterface;
+use Symfony\Component\Security\Core\SecurityContextInterface;
 
 /**
  * Helper for product draft datagrid
@@ -21,6 +23,17 @@ use PimEnterprise\Bundle\WorkflowBundle\Model\ProductDraft;
  */
 class GridHelper
 {
+    /** @var SecurityContextInterface  */
+    protected $securityContext;
+
+    /**
+     * @param SecurityContextInterface $securityContext
+     */
+    public function __construct(SecurityContextInterface $securityContext)
+    {
+        $this->securityContext = $securityContext;
+    }
+
     /**
      * Returns callback that will disable approve and refuse buttons
      * given product draft status
@@ -30,11 +43,17 @@ class GridHelper
     public function getActionConfigurationClosure()
     {
         return function (ResultRecordInterface $record) {
+            if (null !== $this->securityContext &&
+                false === $this->securityContext->isGranted(Attributes::EDIT_ATTRIBUTES, $record->getRootEntity())
+            ) {
+                return ['approve' => false, 'refuse' => false, 'remove' => false];
+            }
+
             if (ProductDraft::IN_PROGRESS === $record->getValue('status')) {
                 return ['approve' => false, 'refuse' => false];
-            } else {
-                return ['remove' => false];
             }
+
+            return ['remove' => false];
         };
     }